# docs/swagger.yaml
basePath: /api
definitions:
...
  utils.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  utils.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/utils.PaginationData'
    type: object
  utils.PaginationData:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  venue.AutoTimeSlotInput:
    properties:
      booking_type:
        type: string
      court_numbers:
        items:
          type: integer
        type: array
      days_of_week:
        items:
          type: string
        type: array
      duration:
        minimum: 15
        type: integer
      end_date:
        type: string
      end_time:
        type: string
      equipment:
        type: string
      price:
        minimum: 0
        type: number
      start_date:
        type: string
      start_time:
        type: string
    required:
    - court_numbers
    - days_of_week
    - duration
    - end_date
    - end_time
    - price
    - start_date
    - start_time
    type: object
  venue.Booking:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      end_time:
        type: string
      ground:
        $ref: '#/definitions/venue.Ground'
      ground_id:
        type: integer
      id:
        type: integer
      purpose:
        type: string
      start_time:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  venue.CourtInput:
    properties:
      description:
        type: string
      name:
        type: string
      type:
        type: string
    required:
    - name
    - type
    type: object
  venue.CreateBookingRequest:
    properties:
      end_time:
        type: string
      ground_id:
        type: integer
      purpose:
        type: string
      start_time:
        type: string
    required:
    - end_time
    - ground_id
    - start_time
    type: object
  venue.Ground:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
      venue_id:
        type: integer
    type: object
  venue.TimeSlot:
    properties:
      booked_by:
        type: integer
      booking_type:
        type: string
      court_number:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      end_time:
        type: string
      equipment:
        type: string
      id:
        type: integer
      is_booked:
        type: boolean
      price:
        type: number
      start_time:
        type: string
      updated_at:
        type: string
      venue_id:
        type: integer
    type: object
  venue.TimeSlotInput:
    properties:
      booking_type:
        type: string
      court_number:
        minimum: 1
        type: integer
      end_time:
        type: string
      equipment:
        type: string
      price:
        minimum: 0
        type: number
      start_time:
        type: string
    required:
    - court_number
    - end_time
    - price
    - start_time
    type: object
...
  venue.Venue:
    properties:
      available:
        type: boolean
      capacity:
        type: integer
      contact_info:
        type: string
      coordinates:
        type: string
      court_count:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      facilities:
        type: string
      hourly_rate:
        type: number
      id:
        type: integer
      images:
        type: string
      location:
        type: string
      manager_id:
        type: integer
      name:
        type: string
      social_hours:
        type: string
      updated_at:
        type: string
    type: object
  venue.VenueInput:
    properties:
      available:
        type: boolean
      capacity:
        type: integer
      contact_info:
        type: string
      coordinates:
        type: string
      court_count:
        minimum: 1
        type: integer
      description:
        type: string
      facilities:
        type: string
      hourly_rate:
        minimum: 0
        type: number
      images:
        type: string
      location:
        type: string
      name:
        type: string
      social_hours:
        type: string
    required:
    - court_count
    - hourly_rate
    - location
    - name
    type: object
host: localhost:8088
info:
  contact: {}
  description: "This is a  server for Sport_go\U0001F3CF."
  title: MiowNation REST API(-_-)
  version: "1.0"
paths:
  /api/bookings:
    get:
      consumes:
      - application/json
      description: Retrieves all bookings made by the current user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        maximum: 100
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of user's bookings
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get user's bookings
      tags:
      - bookings
    post:
      consumes:
      - application/json
      description: Creates a new booking for a specific ground/court
      parameters:
      - description: Booking details
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/venue.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Booking created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Ground not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Time slot not available
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new booking
      tags:
      - bookings
  /api/bookings/{booking_id}:
    delete:
      consumes:
      - application/json
      description: Cancels a specific booking and releases the time slot
      parameters:
      - description: Booking ID
        in: path
        name: booking_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Booking cancelled successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Booking not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Cannot cancel booking
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Cancel a booking
      tags:
      - bookings
    get:
      consumes:
      - application/json
      description: Retrieves details of a specific booking
      parameters:
      - description: Booking ID
        in: path
        name: booking_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Booking details
          schema:
            $ref: '#/definitions/venue.Booking'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Booking not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get booking details
      tags:
      - bookings
...
  /manager/venues:
    post:
      consumes:
      - application/json
      description: Create a new venue with the provided details
      parameters:
      - description: Venue information
        in: body
        name: venue
        required: true
        schema:
          $ref: '#/definitions/venue.VenueInput'
      produces:
      - application/json
      responses:
        "201":
          description: Venue created successfully
          schema:
            $ref: '#/definitions/venue.Venue'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new venue
      tags:
      - venues
  /manager/venues/{venue_id}:
    delete:
      description: Delete an existing venue and all its associated data
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Venue deleted successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Invalid venue ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden - not the venue manager
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Venue not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete venue
      tags:
      - venues
    put:
      consumes:
      - application/json
      description: Update an existing venue's details
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: integer
      - description: Updated venue information
        in: body
        name: venue
        required: true
        schema:
          $ref: '#/definitions/venue.VenueInput'
      produces:
      - application/json
      responses:
        "200":
          description: Venue updated successfully
          schema:
            $ref: '#/definitions/venue.Venue'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden - not the venue manager
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Venue not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - Bearer: []
      summary: Update venue
      tags:
      - venues
  /manager/venues/{venue_id}/courts:
    post:
      consumes:
      - application/json
      description: Add a new court to an existing venue
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: integer
      - description: Court information
        in: body
        name: court
        required: true
        schema:
          $ref: '#/definitions/venue.CourtInput'
      produces:
      - application/json
      responses:
        "201":
          description: Court added successfully
          schema:
            $ref: '#/definitions/venue.Ground'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden - not the venue manager
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Venue not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - Bearer: []
      summary: Add court to venue
      tags:
      - venues
  /manager/venues/{venue_id}/courts/{court_id}:
    delete:
      description: Delete an existing court from a venue
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: integer
      - description: Court ID
        in: path
        name: court_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Court deleted successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Invalid input or court doesn't belong to venue
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden - not the venue manager
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Court or venue not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete court
      tags:
      - venues
    put:
      consumes:
      - application/json
      description: Update an existing court's details
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: integer
      - description: Court ID
        in: path
        name: court_id
        required: true
        type: integer
      - description: Updated court information
        in: body
        name: court
        required: true
        schema:
          $ref: '#/definitions/venue.CourtInput'
      produces:
      - application/json
      responses:
        "200":
          description: Court updated successfully
          schema:
            $ref: '#/definitions/venue.Ground'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden - not the venue manager
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Court or venue not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - Bearer: []
      summary: Update court
      tags:
      - venues
  /manager/venues/{venue_id}/timeslots:
    post:
      consumes:
      - application/json
      description: Create one or more time slots for a venue
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: integer
      - description: Time slot information
        in: body
        name: timeSlots
        required: true
        schema:
          items:
            $ref: '#/definitions/venue.TimeSlotInput'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Time slots created successfully
          schema:
            items:
              $ref: '#/definitions/venue.TimeSlot'
            type: array
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden - not the venue manager
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Venue not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Time slot conflict
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - Bearer: []
      summary: Create time slots
      tags:
      - venues
  /manager/venues/{venue_id}/timeslots/{timeslot_id}:
    delete:
      description: Delete an existing time slot
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: integer
      - description: Time Slot ID
        in: path
        name: timeslot_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Time slot deleted successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Invalid input or time slot doesn't belong to venue
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden - not the venue manager
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Time slot or venue not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Cannot delete a booked time slot
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete time slot
      tags:
      - venues
    put:
      consumes:
      - application/json
      description: Update an existing time slot's details
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: integer
      - description: Time Slot ID
        in: path
        name: timeslot_id
        required: true
        type: integer
      - description: Updated time slot information
        in: body
        name: timeSlot
        required: true
        schema:
          $ref: '#/definitions/venue.TimeSlotInput'
      produces:
      - application/json
      responses:
        "200":
          description: Time slot updated successfully
          schema:
            $ref: '#/definitions/venue.TimeSlot'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden - not the venue manager
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Time slot or venue not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Time slot conflict
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - Bearer: []
      summary: Update time slot
      tags:
      - venues
  /manager/venues/{venue_id}/timeslots/auto:
    post:
      consumes:
      - application/json
      description: Generate time slots automatically for a venue based on specified
        parameters
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: integer
      - description: Auto time slot generation parameters
        in: body
        name: autoSlots
        required: true
        schema:
          $ref: '#/definitions/venue.AutoTimeSlotInput'
      produces:
      - application/json
      responses:
        "201":
          description: Time slots created successfully
          schema:
            items:
              $ref: '#/definitions/venue.TimeSlot'
            type: array
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden - not the venue manager
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Venue not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Time slot conflict
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - Bearer: []
      summary: Generate time slots automatically
      tags:
      - venues
...
  /venues:
    get:
      description: Get a paginated list of all venues with optional filters
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: Filter by availability
        in: query
        name: available
        type: boolean
      - description: Filter by location (partial match)
        in: query
        name: location
        type: string
      - description: Filter by minimum number of courts
        in: query
        name: min_courts
        type: integer
      - description: Filter by maximum hourly rate
        in: query
        name: max_price
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: List of venues
          schema:
            allOf:
            - $ref: '#/definitions/utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/venue.Venue'
                  type: array
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get all venues
      tags:
      - venues
  /venues/{venue_id}:
    get:
      description: Get detailed information about a venue by its ID
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Venue details
          schema:
            $ref: '#/definitions/venue.Venue'
        "400":
          description: Invalid venue ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Venue not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get venue by ID
      tags:
      - venues
  /venues/{venue_id}/courts:
    get:
      description: Get all courts for a specific venue
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of courts
          schema:
            items:
              $ref: '#/definitions/venue.Ground'
            type: array
        "400":
          description: Invalid venue ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Venue not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get venue courts
      tags:
      - venues
  /venues/{venue_id}/timeslots:
    get:
      description: Get time slots for a specific venue, optionally filtered by date
        and court number
      parameters:
      - description: Venue ID
        in: path
        name: venue_id
        required: true
        type: integer
      - description: Filter by date (YYYY-MM-DD format)
        in: query
        name: date
        type: string
      - description: Filter by court number
        in: query
        name: court_number
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of time slots
          schema:
            items:
              $ref: '#/definitions/venue.TimeSlot'
            type: array
        "400":
          description: Invalid input parameters
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Venue not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get venue time slots
      tags:
      - venues
swagger: "2.0"